\name{prep_shem_col}
\alias{prep_shem_col}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prepare la table de donnees
}
\description{
Prepare la table de donnees en standardisant les noms des colonnes et en passant la date en POSIXct.
}
\usage{
prep_shem_col()
}
%- maybe also 'usage' for other objects documented here.
\details{
RAS
}
\value{
Retourne la table (dataframe) clean
}
\references{
RAS
}
\author{
Corentin Bui <corentin.bui@shem.engie.com>
}
\note{
RAS
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
RAS
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function ()
{
    return("Hello !")
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
